Update 4
begin avalaibility

+++ src/main/java/Presentation/ViewTracksFromPlaylistController.java
@@ -19,11 +19,10 @@ import java.util.List;
  */
 @WebServlet("/ViewTracksFromPlaylist")
 public class ViewTracksFromPlaylistController extends HttpServlet {
-
+    List<TrackModel> trackmodels = new ArrayList<TrackModel>();
     private TrackModel tm = new TrackModel();
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        List<TrackModel> trackmodels = new ArrayList<TrackModel>();
         HttpSession s =request.getSession();
         String playlist = (String)s.getAttribute("playlistname");
         trackmodels =tm.getTracksFromPlaylist(playlist);
@@ -34,11 +33,17 @@ public class ViewTracksFromPlaylistController extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
         HttpSession s =request.getSession();
         String playlist = (String)s.getAttribute("playlistname");
+        s.setAttribute("playlist", trackmodels);
         String title = request.getParameter("title");
         String performer = request.getParameter("performer");
-        tm = new TrackModel();
-        tm.deleteTrackFromPlaylist(title, performer, playlist);
-        response.sendRedirect("ViewTracksFromPlaylist");
+        if(request.getParameter("DeleteTrack") != null) {
+            tm.deleteTrackFromPlaylist(title, performer, playlist);
+            response.sendRedirect("ViewTracksFromPlaylist");
+        }
+        else if(request.getParameter("changeAvailability") != null) {
+            response.sendRedirect("ChangeAvailability");
+        }
+
 
     }
 }
